(Refactoring) Kept metadata in models.__.__init__, moved get_iterator_and_processor in models.___.imports
