+++ pie_extended/models/fro/__init__.py
@@ -1,6 +1,4 @@
 from ...utils import Metadata, File, get_path
-from .get import get_iterator_and_processor
-from ...pipeline.iterators.proto import DataIterator
 
 DESC = Metadata(
     "Ancien Francais (ENC)",
