+++ pie_extended/models/lasla/__init__.py
@@ -1,2 +1,19 @@
-from pie_extended.models.lasla.consts import DOWNLOADS, Models, Disambiguator, addons, DESC
-from pie_extended.models.lasla.get import get_iterator_and_processor
+from ...utils import Metadata, File, get_path
+
+
+DESC = Metadata(
+    "LASLA-ENC",
+    "lat",
+    ["Thibault Clérice"],
+    "Model trained on LASLA data without disambiguation",
+    "https://github.com/chartes/deucalion-model-lasla"
+)
+
+DOWNLOADS = [
+    File("https://github.com/chartes/deucalion-model-lasla/raw/master/latin-straight.json", "latin-straight.json"),
+    File("https://github.com/chartes/deucalion-model-lasla/raw/master/latin-pos.json", "latin-pos.json"),
+    File("https://github.com/chartes/deucalion-model-lasla/raw/master/latin-needs.json", "latin-needs.json"),
+    File("https://github.com/chartes/deucalion-model-lasla/raw/master/model.tar", "model.tar")
+]
+
+Models = "<{},lemma,Voice,Mood,Deg,Numb,Person,Tense,Case,Gend,pos>".format(get_path("lasla", "model.tar"))
