+++ tests/test_models/test_fro.py
@@ -1,4 +1,4 @@
-from pie_extended.models.fro.get import get_iterator_and_processor
+from pie_extended.models.fro.imports import get_iterator_and_processor
 from pie_extended.testing_utils import FakeTagger
 from typing import List, Tuple
